Disable archive of Odyssey NOR image into SBE debug tools for lab config
Provided a flag --disable_arch_nor_img to control archive from command
line as well

+++ imageBuild/imageBuild.py
@@ -438,6 +438,8 @@ parser.add_argument('--allowToSign', action='store_true',
                          'configured image_sections.')
 parser.add_argument('--no_downloads', action='store_true',
                     help='Disable downloading any repositories/binaries etc.')
+parser.add_argument('--disable_arch_nor_img', action='store_true',
+                    help='disable nor image copy into debug archive')
 args = parser.parse_args()
 
 # process the configuration file and load needed modules whos location is based on
@@ -829,32 +831,34 @@ if concatCopies > 1:
 
     f1.close()
 
-    # Copy odyssey_nor_DD1.img into odyssey_sbe_debug_DD1.tar.gz
-    archSbeDebugTar = os.path.join(sbeImageDir, "odyssey/odyssey_sbe_debug_DD1.tar.gz")
-    if not os.path.exists(archSbeDebugTar):
-        print(f"{archSbeDebugTar} does not exist", file=sys.stderr)
-        sys.exit(1)
-    else:
-        print("INFO: Untar odyssey_sbe_debug_DD1.tar.gz")
-        archive = tarfile.open(archSbeDebugTar)
-        pathSbeDebugTar = os.path.dirname(os.path.abspath(archSbeDebugTar))
-        archive.extractall(pathSbeDebugTar)
-        archive.close()
+    if not args.disable_arch_nor_img and "lab_image_config" not in args.configfile:
+        print("INFO: Odyssey pnor image config")
+        # Copy odyssey_nor_DD1.img into odyssey_sbe_debug_DD1.tar.gz
+        archSbeDebugTar = os.path.join(sbeImageDir, "odyssey/odyssey_sbe_debug_DD1.tar.gz")
+        if not os.path.exists(archSbeDebugTar):
+            print(f"{archSbeDebugTar} does not exist", file=sys.stderr)
+            sys.exit(1)
+        else:
+            print("INFO: Untar odyssey_sbe_debug_DD1.tar.gz")
+            archive = tarfile.open(archSbeDebugTar)
+            pathSbeDebugTar = os.path.dirname(os.path.abspath(archSbeDebugTar))
+            archive.extractall(pathSbeDebugTar)
+            archive.close()
 
-        # Remove odyssey_sbe_debug_DD1.tar.gz
-        os.remove(archSbeDebugTar)
+            # Remove odyssey_sbe_debug_DD1.tar.gz
+            os.remove(archSbeDebugTar)
 
-        print("INFO: Copy odyssey_nor_DD1.img into extracted odyssey_debug_files_tools")
-        pathSbeDebugTools = os.path.join(pathSbeDebugTar, "odyssey_debug_files_tools")
-        shutil.copy(imagefile, pathSbeDebugTools)
+            print("INFO: Copy odyssey_nor_DD1.img into extracted odyssey_debug_files_tools")
+            pathSbeDebugTools = os.path.join(pathSbeDebugTar, "odyssey_debug_files_tools")
+            shutil.copy(imagefile, pathSbeDebugTools)
 
-        print("INFO: Archive odyssey_debug_files_tools into tar file odyssey_sbe_debug_DD1.tar.gz")
-        archive = tarfile.open(archSbeDebugTar, "w:gz")
-        archive.add(pathSbeDebugTools, arcname=os.path.basename(pathSbeDebugTools))
-        archive.close()
+            print("INFO: Archive odyssey_debug_files_tools into tar file odyssey_sbe_debug_DD1.tar.gz")
+            archive = tarfile.open(archSbeDebugTar, "w:gz")
+            archive.add(pathSbeDebugTools, arcname=os.path.basename(pathSbeDebugTools))
+            archive.close()
 
-        # Remove directory odyssey_debug_files_tools
-        shutil.rmtree(pathSbeDebugTools)
+            # Remove directory odyssey_debug_files_tools
+            shutil.rmtree(pathSbeDebugTools)
 
 #--------------------------
 # ecc
